requirements.txt
fastapi
uvicorn
openai
azure-search-documents
pandas
python-dotenv
requests
